+++ .prettierrc.js
@@ -1,1 +1,1 @@
-module.exports = require('@muravjev/prettier-config-monorepo');
+module.exports = require('@muravjev/configs-prettier');